Fixed problems with test execution in app project.
In order to be executed, tests must use a special Android test runner.
The test classes have been rewritten accordingly.

The build was failing when compiling the release unit tests because
the dependency to fragment testing could not be resolved. Therefore,
the test compilation task has been disabled for now.