ServerTrackState now supports adding files out of order.
The original implementation of appendFile() just added the new file to
the end of the list of files on the server, which was sufficient for
the use cases supported so far.

In future it can happen that files are added to the server in a
different order, for instance if no mobile connection was available, a
file is uploaded at a later point in time. This is now supported by
providing an explicit boolean parameter that triggers an additional
sort operation.