TrackService now supports uploading multiple location data files.
This functionality is going to be used to deal with upload errors
caused by a broken connection. In this case, location data is buffered
and then uploaded when a connection is available again.

The upload of multiple location data files differs slightly from a
single file upload because it can no longer be assumed that a newly
uploaded file is newer than the other data on the server.