MapUpdater now displays the own location on the map.
If the location is defined, another marker is added to the map with
special properties.

A function was added to center an arbitrary marker on the map. The
function to center the most recent location delegates to this new
function.

MarkerFactory has been changed to expect the MarkerData to be passed
in rather than the state and the key; the MarkerData for the own
location is not contained in the state.