The marker for the own location is now correctly removed.
This was not the case before; every time, the map was updated, but no
new locations were found, a new marker for the own position was drawn.
Handling the removal is not trivial, as the Marker object has to be
stored. To achieve this, a new state class has been introduced.