Upgraded some dependencies as recommended by the IDE.
After the upgrade, an error was reported in LocationRetriever because
of a missing check for the right to request the user location. As a
quick fix, a SecurityException is now handled. As a future
improvement, the correct workflow to request a missing right should be
implemented.

Also updated dependency to Ok Http client; unfortunately, it was not
possible to switch to the most recent version, as this version could
not be resolved.