Added two more properties to TrackConfig.
The properties have the purpose to detect inaccurate GPS locations.
The idea is to calculate the speed between two GPS locations in
sequence retrieved by LocationRetriever. If there is a sudden,
unrealistic speed increase, the latest GPS update is probably wrong.

The new properties support the customization of this algorithm.