LocationRetrieverImpl: Changed error handling.
According to the annotations in LocationCallback(), it does not seem
to be possible that onLocationResult() is called with a *null* result.
Therefore, a non-nullable type can be used here, and the corresponding
test case can be dropped.

In the test class, a number of warnings regarding nullability could be
fixed. One warning about a null value passed for the Looper parameter
when calling requestLocationUpdates() remains. According to the
documentation, the value null is allowed and references the current
thread.