MapFragment: Implemented official request for location permission.
In newer Android versions, apps should ask the user to grant
permissions when needed for the first time. This concept has now been
implemented as a PoC for the location permission in MapFragment.

As this logic is also needed for the tracking fragment, it is going to
be extracted into a helper class.

Some test cases for MapFragment needed to be disabled, as they were
broken by the new logic.