Build: Cleaned up test dependencies; made UI tests run again.
The tests based on Robolectric could not be executed when in the test
options the useJUnitPlatform() property was set; running such a test
yields the message "no test events received".

This was fixed by adding a dependency to the JUnit vintage engine.
Unneeded dependencies were removed. Robolectric was upgraded to
version 4.8.