PreferencesHandler: Made conversion functions public.
The functions to convert property values to Ints, Doubles, or Dates
are now public. In the future, PreferencesHandler will serve as a thin
wrapper over the application's shared preferences, and configuration
classes access their own properties. Therefore, the functions must be
accessible from the outside.

Converted the test class from a StringSpec to a WordSpec to have more
structure in the tests for different functionality.