TrackConfig: Added functionality to create itself from preferences.
The PreferencesHandler class is currently doing too much. It has
knowledge about structures from different areas of the application.

As a first step to remedy this, the creation of a TrackConfig can now
be done by TrackConfig itself. The logic in PreferencesHandler will be
removed after the call sites have been adapted.

Signed-off-by: oheger <oliver.heger@oliver-heger.de>