TrackViewModel: Extracted an interface.
The interface is not actually needed for the production code, but for
preview composable functions. The preview cannot be rendered when it
accesses the real view model. With the interface, it is now possible
to construct a fake model and pass it to a function that requires a
model.

TrackUi.kt makes use of the new interface. The preview function can
now display the whole track view.