TrackConfig: Changed handling of default values.
Configuration default values are not longer defined using constants,
but there is now a default configuration instance holding these
values. The function to initialize defaults in the companion object
has been replaced by an instance method to save the current values,
optionally keeping existing values. This gives more flexibility;
saving of configuration values is going to be needed in the future
when the configuration UI is reworked.