PreferencesHandler: Removed further obsolete functionality.
The creation of the ServerConfig is no longer needed.

Note: Currently, the class still contains the check for properties of
this configuration, which is needed to stop tracking if there is a
configuration change. This whole mechanism is going to change in the
future: It will then only be possible to modify the configuration if
tracking is disabled.