ConfigUi: ConfigItem now supports different editor types.
The function has been generalized to provide the framework around a
config item to be edited. It now expects a composable function as
argument that generates the actual editor. The value to be edited can
be of a different type.

The handling of the remembered value was a bit problematic which could
lead to data loss when the Save button was clicked without changing
the text field. This has been fixed.

Signed-off-by: oheger <oliver.heger@oliver-heger.de>