Config: Added a model class for a duration editor.
The idea behind this model is that a duration can be split into
multiple components, such as seconds, minutes, hours, etc. Each
component can be edited in a single numeric field.

The model manages the single components. A duration in seconds can be
split into these components, and from the components, a duration can
be computed.

Signed-off-by: oheger <oliver.heger@oliver-heger.de>