Build: Updated Compose version from 1.1.1 to 1.2.1.
This caused test failures in TrackUiSpec, since it is no longer
possible to change the content of the test rule. The failures could be
fixed by splitting the test class into two: One based on
createAndroidComposeRule() that tests the view integration in the
application, and one based on a test rule whose content is set
dynamically.

Signed-off-by: oheger <oliver.heger@oliver-heger.de>