ConfigUi: Reworked DurationEditorModel.
The representation of a duration based on its single components is
useful for other use cases, too, not only when editing durations in
the configuration UI. Therefore, the DurationEditorModel class has
been renamed to DurationModel, and UI-specific aspects have been
removed.

The Saver is now provided in a separate source file, together with a
convenience function to remember a duration.

The new DurationModel class is going to be used to format durations in
a later commit.

Signed-off-by: oheger <oliver.heger@oliver-heger.de>