TimeDeltaFormatter: Reworked class.
The class has been moved to the duration package. It now uses
DurationModel to obtain the single components of the duration to be
formatted.

The create() function was removed, since it introduced a dependency to
UI classes; the unit strings must now be passed in when an instance is
created.

Signed-off-by: oheger <oliver.heger@oliver-heger.de>