Receiver UI: Added a new MapStateLoader class.
The class provides a subset of the functionality offered by
MapUpdater: It can retrieve data from the server, but it does not
update a map. In Compose, a declarative approach is used to add
markers to a map view; map objects are not manipulated directly any
more. The new class is going to be used in this context.

Signed-off-by: oheger <oliver.heger@oliver-heger.de>