Config: Added a new ConfigManager class.
The class manages the different types of configuration objects
supported by the application. It loads the settings from shared
preferences and also allows updates.

The main purpose of this class is to allow view models to monitor
configurations for updates, since this might require some actions to
be taken.

Strangely, after adding this new class and its tests,
TrackConfigSpec was failing. This could be fixed by calling
unmockAll() explicitly before each test. (It was not sufficient to
call this function only once before the spec.)

Signed-off-by: oheger <oliver.heger@oliver-heger.de>