TrackStatsFormatter: Added a property for a default instance.
The formatter is used by multiple UIs that are served from different
view models. Therefore, it does not make sense to make it available
via a property of the view model. Via the new INSTANCE property, the
default formatter can now be accessed everywhere.

The formatter property was removed from TrackViewModel.

Signed-off-by: oheger <oliver.heger@oliver-heger.de>