PreferencesHandler: Reworked access to properties.
Introduced functions to get properties of other data types, so that
the class is now a wrapper over the SharedPreferences API. This
simplifies clients, since all types can now be handled in the same
way.

Made the wrapped SharedPreferences object private; it should not be
accessed any more from clients.

The functions to query Int and Double properties no longer try to read
strings from the preferences. This is no longer necessary since
configuration objects are stored directly and not via the
configuration framework. This caused TrackConfigUiSpec to fail, which
has been temporarily disabled.

Signed-off-by: oheger <oliver.heger@oliver-heger.de>