ReceiverViewModel: Introduced actions.
The model can now handle actions triggered by the UI that are defined
by a new enum class.

The idea is that these actions correspond to buttons on the receiver
UI. When the user clicks such a button, the function is called,
causing some update of the receiver state.

Currently, an action to reload the location state from the server
immediately has been implemented.

Signed-off-by: oheger <oliver.heger@oliver-heger.de>