Navigation: Introduced Compose navigation for the sender screen.
Added a new main screen to represent the whole application, including
a top bar with a drawer that allows navigating to the different
supported screens. This screen is currently displayed by
TrackFragment (which looks a bit strange). This way, the single
screens of the application can be reworked one by one. When this is
done, the main activity can be changed to call the top-level
Composable function, and all fragments can be removed.

Signed-off-by: oheger <oliver.heger@oliver-heger.de>