Clean-up: Removed obsolete ComposeFragment class.
This class was used as a base class for fragments during the
transition to Compose. After this is done and no fragments are used
any longer, it can be removed.

With this class being gone, the view binding feature is no longer
needed, and the corresponding setting can be removed from the app's
build file.

Signed-off-by: oheger <oliver.heger@oliver-heger.de>